cloud_functions/zoom_to_transcript/requirements.txt
functions-framework==3.*
google-cloud-speech==2.*
google-cloud-storage==2.*
requests==2.*
pydub==0.25.*
